requirements.txt
Flask==1.1.2
requests-oauthlib==1.3.0
python-dotenv==0.15.0
google-cloud-vision==2.0.0
google-cloud-datastore==2.1.0
google-cloud-storage==1.35.0
pymongo==3.11.2
dnspython==2.1.0